backend/ia-seguridad/requirements.txt
flask==3.0.0
flask-cors==4.0.0
docker==7.0.0
google-generativeai==0.8.3
openai==1.6.1
requests==2.31.0
pymysql==1.1.0
sqlalchemy==2.0.23
redis==5.0.1
python-dotenv==1.0.0
schedule==1.2.0
psutil==5.9.6
